refactor: centralize consolidation utilities
