+++ src/core/consolidation/__init__.py
@@ -1,0 +1,28 @@
+"""Single source of truth for consolidation utilities."""
+
+from .base import ConsolidationBase
+from .utils import (
+    ConfigUtils,
+    FileSystemUtils,
+    IOUtils,
+    LoggingUtils,
+    MathUtils,
+    StringUtils,
+    SystemUtils,
+    ValidationUtils,
+)
+from .utils_system import ConsolidatedUtilsSystem, consolidated_utils
+
+__all__ = [
+    "ConsolidationBase",
+    "ConsolidatedUtilsSystem",
+    "consolidated_utils",
+    "ConfigUtils",
+    "FileSystemUtils",
+    "IOUtils",
+    "LoggingUtils",
+    "MathUtils",
+    "StringUtils",
+    "SystemUtils",
+    "ValidationUtils",
+]
